test: split agent tests (mitigate juju bug)
